+++ app/Services/Interfaces/IProductService.php
@@ -6,7 +6,9 @@ interface IProductService
 {
     public function getAll();
     public function getById(int $id);
+    public function getByTitle(string $title);
     public function insert(array $data);
     public function update(int $id, array $data);
     public function delete(int $id);
+    public function getProductCategoryIds(int $productId);
 }