+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # FWS test task CHANGELOG
 
+### 2024. 10. 25.
+### @tbence
+- Added CSV import process
+- Added chunk based CSV import
+- Implemented Redis for job cache
+
 ### 2024. 10. 24.
 #### @tbence
 - Added new tables: `categories`, `products`, `products_categories`